perl-Locale-Gettext: work around t/frconvert.t failure
Since the upgrade to the MSYS2 runtime v3.1.4, LC_MESSAGES defaults to
`en_US.UTF-8`, which misses the message prepared for that test (which
assumes that `LC_MESSAGES` is set to the empty string.

When forcing `setlocale(LC_MESSAGES, "")` just before the `dgettext()`
call in the native code, it still works, but sadly, the same call in the
Perl code of `t/frconvert.t` seems not to fix the issue.

So for the time being, just skip the test by setting `LC_MESSAGES=C`
(which `test_data/gen_test_data.pl` does not like and therefore causes
the script to force the test to be skipped).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>