openssh: backport patch to fix "invalid format"
Due to a bug in newer OpenSSH versions, when the public key is absent,
loading a private key in OpenSSL format will print the warning "invalid
format".

This has been fixed in OpenSSH, but not released yet. So we simply
backport that fix.

This fixes https://github.com/git-for-windows/git/issues/2743

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>