msys2-runtime: ensure that it is checked out with core.autocrlf=false
Enough scripts in Cygwin rely on the fact that the files are checked out
using Unix line endings to make it unpractical to check them out using
`core.autocrlf=true`.

One symptom encountered in the Git for Windows project is that the
`gentls_offsets` script fails to remove the `#pragma once` line from
`cygtls.h` while copy-editing it into `/tmp/$$.cc`, and therefore the
build will fail to (re-)build `tlsoffsets64.h` (or `tlsoffsets.h` in the
i686 build) with an error message like this:

	/tmp/21341.cc:18:9: error: #pragma once in main file [-Werror]
	   18 | #pragma once
	      |         ^~~~
	cc1plus: all warnings being treated as errors
	g++: error: /tmp/21341-1.cc: No such file or directory
	g++: fatal error: no input files
	compilation terminated.

The proper way to fix this would be to edit the `.gitattributes`, of
course. But the way the MSYS2 runtime is actually built is not by
cloning the `msys2-runtime` repository. Instead, Cygwin's repository is
cloned, and the patches are applied one by one on top. Therefore, if we
added a commit to the `msys2-runtime` repository that adjusts the
`.gitattributes`, that commit would then be turned into a patch in
`MSYS2-packages`' `msys2-runtime/` subdirectory, and it would be only
applied _after_ the repository was cloned, i.e. too late.

So let's just force the source code to be checked out with
`core.autocrlf=false`.

Note: we do _not_ add this to the `core.symlinks` clause because it
needs to be handled slightly differently: after changing
`core.autocrlf`, Git will not realize that it needs to update _all the
files_. Therefore we have to remove the Git index in that case (a bit
heavy-handed, to be sure, but it is the easiest way to force the line
endings to be changed).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>