curl: stop using metalink
The metalink feature was removed in cURL v7.78.0 due to security
concerns.

Git for Windows never used it, it was a feature designed to pick the
best mirror for a given website, and there is no support for it in Git.
It only affects `curl.exe`. For more details, see
https://github.com/curl/curl/blob/70233523976f~2/docs/cmdline-opts/metalink.d

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>